Initial
Working: Grids exist, and the big one will respond to what zone the
"hand" is in.

To do: Actually add the leap stuff, insert audio files so that it gives
audio feedback, and add the timer for response time. Oh, also figure out
how to save those RT's  in a col, then output that col as a .txt.